Move global var offload tests into data_offload tests
