+++ transformations/tests/test_data_offload.py
@@ -5,16 +5,22 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
+from pathlib import Path
 from shutil import rmtree
 import pytest
 
 from loki import (
     Sourcefile, FindNodes, Pragma, PragmaRegion, Loop,
     CallStatement, pragma_regions_attached, get_pragma_parameters,
-    gettempdir, Scheduler, OMNI
+    gettempdir, Scheduler, OMNI, Import
 )
 from conftest import available_frontends
 from transformations import DataOffloadTransformation, GlobalVariableAnalysis, NewGlobalVarOffloadTransformation
+
+
+@pytest.fixture(scope='module', name='here')
+def fixture_here():
+    return Path(__file__).parent
 
 
 @pytest.fixture(name='config')
@@ -497,3 +503,130 @@ def test_global_variable_offload(frontend, key, config, global_variable_analysis
             for pragma in acc_pragmas
         }
         assert variables == expected_declarations[name]
+
+
+@pytest.mark.parametrize('frontend', available_frontends())
+def test_transformation_global_var_import(here, config, frontend):
+    """
+    Test the generation of offload instructions of global variable imports.
+    """
+    config['default']['enable_imports'] = True
+    config['routines'] = {
+        'driver': {'role': 'driver'}
+    }
+
+    scheduler = Scheduler(paths=here/'sources/projGlobalVarImports', config=config, frontend=frontend)
+    scheduler.process(transformation=GlobalVariableAnalysis())
+    scheduler.process(transformation=NewGlobalVarOffloadTransformation())
+
+    driver = scheduler['#driver'].routine
+    moduleA = scheduler['modulea#var0'].scope
+    moduleB = scheduler['moduleb#var2'].scope
+    moduleC = scheduler['modulec#var4'].scope
+
+    # check that global variables have been added to driver symbol table
+    imports = FindNodes(Import).visit(driver.spec)
+    assert len(imports) == 2
+    assert imports[0].module != imports[1].module
+    assert imports[0].symbols != imports[1].symbols
+    for i in imports:
+        assert len(i.symbols) == 2
+        assert i.module.lower() in ('moduleb', 'modulec')
+        assert set(s.name for s in i.symbols) in ({'var2', 'var3'}, {'var4', 'var5'})
+
+    # check that existing acc pragmas have not been stripped and update device/update self added correctly
+    pragmas = FindNodes(Pragma).visit(driver.body)
+    assert len(pragmas) == 4
+    assert all(p.keyword.lower() == 'acc' for p in pragmas)
+
+    assert 'update device' in pragmas[0].content
+    assert 'var2' in pragmas[0].content
+    assert 'var3' in pragmas[0].content
+
+    assert pragmas[1].content == 'serial'
+    assert pragmas[2].content == 'end serial'
+
+    assert 'update self' in pragmas[3].content
+    assert 'var4' in pragmas[3].content
+    assert 'var5' in pragmas[3].content
+
+    # check that no declarations have been added for parameters
+    pragmas = FindNodes(Pragma).visit(moduleA.spec)
+    assert not pragmas
+
+    # check for device-side declarations where appropriate
+    pragmas = FindNodes(Pragma).visit(moduleB.spec)
+    assert len(pragmas) == 2
+    assert pragmas[0].content != pragmas[1].content
+    assert all(p.keyword == 'acc' for p in pragmas)
+    assert all('declare create' in p.content for p in pragmas)
+    assert any('var2' in p.content for p in pragmas)
+    assert any('var3' in p.content for p in pragmas)
+
+    pragmas = FindNodes(Pragma).visit(moduleC.spec)
+    assert len(pragmas) == 2
+    assert pragmas[0].content != pragmas[1].content
+    assert all(p.keyword == 'acc' for p in pragmas)
+    assert all('declare create' in p.content for p in pragmas)
+    assert any('var4' in p.content for p in pragmas)
+    assert any('var5' in p.content for p in pragmas)
+
+
+@pytest.mark.parametrize('frontend', available_frontends())
+def test_transformation_global_var_import_derived_type(here, config, frontend):
+    """
+    Test the generation of offload instructions of derived-type global variable imports.
+    """
+
+    config['default']['enable_imports'] = True
+    config['routines'] = {
+        'driver_derived_type': {'role': 'driver'}
+    }
+
+    scheduler = Scheduler(paths=here/'sources/projGlobalVarImports', config=config, frontend=frontend)
+    scheduler.process(transformation=GlobalVariableAnalysis())
+    scheduler.process(transformation=NewGlobalVarOffloadTransformation())
+
+    driver = scheduler['#driver_derived_type'].routine
+    module = scheduler['module_derived_type#p'].scope
+
+    # check that global variables have been added to driver symbol table
+    imports = FindNodes(Import).visit(driver.spec)
+    assert len(imports) == 1
+    assert len(imports[0].symbols) == 2
+    assert imports[0].module.lower() == 'module_derived_type'
+    assert set(s.name for s in imports[0].symbols) == {'p', 'p0'}
+
+    # check that existing acc pragmas have not been stripped and update device/update self added correctly
+    pragmas = FindNodes(Pragma).visit(driver.body)
+    assert len(pragmas) == 5
+    assert all(p.keyword.lower() == 'acc' for p in pragmas)
+
+    assert 'enter data copyin' in pragmas[0].content
+    assert 'p0%x' in pragmas[0].content
+    assert 'p0%y' in pragmas[0].content
+    assert 'p0%z' in pragmas[0].content
+    assert 'p%n' in pragmas[0].content
+
+    assert 'enter data create' in pragmas[1].content
+    assert 'p%x' in pragmas[1].content
+    assert 'p%y' in pragmas[1].content
+    assert 'p%z' in pragmas[1].content
+
+    assert pragmas[2].content == 'serial'
+    assert pragmas[3].content == 'end serial'
+
+    assert 'exit data copyout' in pragmas[4].content
+    assert 'p%x' in pragmas[4].content
+    assert 'p%y' in pragmas[4].content
+    assert 'p%z' in pragmas[4].content
+
+    # check for device-side declarations
+    pragmas = FindNodes(Pragma).visit(module.spec)
+    assert len(pragmas) == 3
+    assert all(p.keyword == 'acc' for p in pragmas)
+    assert all('declare create' in p.content for p in pragmas)
+    assert 'p' in pragmas[0].content
+    # Note: g is not offloaded because it is not used by the kernel (albeit imported)
+    assert 'p0' in pragmas[1].content
+    assert 'p_array' in pragmas[2].content
